Optimize algorithms in BackOffNGram
The following optimizations were implemented:

- All possible calls to method A() are precalculated in order to avoid
unnecessary computation.

- When calling self.count(ngram) (or self.A(tokens)) with a ngram (or tokens)
that is not in self.counts (or self.cache_A) a pair (key, default_value) is no
longer created, instead the method just returns the default value (0). This way
unnecessary memory usage can be reduced.

In order for the precalculation of A modifications in the super class were
necessary which in turn implied more modifications in the classes that derived
from that super class.