Implement CKYParser
Note that the complexity is O(n**3 * G) where G is the grammar size. There is
another implementation with a worse time complexity ( O(n**3 * G ** 2) ) that
calculates the cartesian product between the symbols (nonterminals and
terminals) on the bottom right interval and on the bottom left interval of the
current cell and then filters that cartesian product discarding all pairs (B, C)
that do not belong to any production in the grammar.